Remove undelivered modal + move SentToInactiveUserModal into correct folder
